requirements.txt
Django==3.0
djangorestframework==3.9.0
psycopg2>=2.7.5,<2.8.0

flake8==3.7.0

